+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 ## Unreleased
 
+### Changed
+- Show an error message when users try to connect a social account that is already attached to their account.
+
 ### Fixed
 - Fixed a bug where the `site` query param was present in the OAuth redirect URI.